[APIs: analytics for a url and couts datewise apis]

+++ src/main/java/com/url/shortner/repositories/UrlMappingRepository.java
@@ -6,7 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UrlMappingRepository extends JpaRepository<UrlMapping, Integer> {
+public interface UrlMappingRepository extends JpaRepository<UrlMapping, Long> {
     UrlMapping findByShortUrl(String shortUrl);
     List<UrlMapping> findByUser(User user);
+//    List<Cli>
 }
